Refactor duplicate methods to shared functions
