+++ flutter-gui-tests/testSrc/io/flutter/tests/gui/TestSuite.kt
@@ -15,5 +15,6 @@ import org.junit.runners.Suite
 // The log file is at flutter-gui-tests/build/idea-sandbox/system/log/idea.log
 // The test report is at flutter-gui-tests/build/reports/tests/test/index.html
 @RunWith(GuiTestSuiteRunner::class)
-@Suite.SuiteClasses(SmokeTest::class, InspectorTest::class, PerfTest::class)
+@Suite.SuiteClasses(InspectorTest::class)
+//@Suite.SuiteClasses(SmokeTest::class, InspectorTest::class, PerfTest::class)
 class TestSuite : GuiTestSuite()
